# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AdVise API",
    "description": "AdVise backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/predictions/": {
      "get": {
        "tags": [
          "Predictions"
        ],
        "summary": "Get Predictions",
        "operationId": "get_predictions_predictions__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/PredictionResponse"
                  },
                  "type": "object",
                  "title": "Response Get Predictions Predictions  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Predictions"
        ],
        "summary": "Create Prediction",
        "operationId": "create_prediction_predictions__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/predictions/preview": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Preview Prediction",
        "operationId": "preview_prediction_v1_predictions_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/prediction-runs/{run_id}": {
      "get": {
        "tags": [
          "predictions"
        ],
        "summary": "Get Prediction Run",
        "operationId": "get_prediction_run_v1_prediction_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PredictionCreate": {
        "properties": {
          "campaign_id": {
            "type": "integer",
            "title": "Campaign Id"
          },
          "predicted_ctr": {
            "type": "number",
            "title": "Predicted Ctr"
          },
          "predicted_conversion_rate": {
            "type": "number",
            "title": "Predicted Conversion Rate"
          }
        },
        "type": "object",
        "required": [
          "campaign_id",
          "predicted_ctr",
          "predicted_conversion_rate"
        ],
        "title": "PredictionCreate"
      },
...
      "PredictionPreviewRequest": {
        "properties": {
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "campaign_intent": {
            "type": "string",
            "title": "Campaign Intent"
          },
          "product_type": {
            "type": "string",
            "title": "Product Type"
          },
          "cta_type": {
            "type": "string",
            "title": "Cta Type"
          },
          "audience_temperature": {
            "type": "string",
            "title": "Audience Temperature"
          },
          "customer_type": {
            "type": "string",
            "title": "Customer Type"
          },
          "budget": {
            "type": "number",
            "title": "Budget"
          },
          "duration_days": {
            "type": "integer",
            "title": "Duration Days"
          },
          "creative_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creative Count",
            "default": 1
          },
          "audience_age": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audience Age",
            "description": "e.g. 25-34"
          },
          "audience_gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audience Gender",
            "description": "male / female"
          },
          "audience_location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audience Location"
          },
          "audience_interests": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audience Interests"
          },
          "career": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Career"
          },
          "creative_image_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creative Image Base64"
          },
          "creative_asset_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creative Asset Url",
            "description": "Hosted creative URL for ads.creative_url"
          },
          "campaign_id": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id"
          },
          "ad_id": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad Id"
          }
        },
        "type": "object",
        "required": [
          "platform",
          "campaign_intent",
          "product_type",
          "cta_type",
          "audience_temperature",
          "customer_type",
          "budget",
          "duration_days"
        ],
        "title": "PredictionPreviewRequest"
      },
      "PredictionPreviewResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "campaign_intent": {
            "type": "string",
            "title": "Campaign Intent"
          },
          "target_metric": {
            "type": "string",
            "title": "Target Metric"
          },
          "target_label": {
            "type": "string",
            "title": "Target Label"
          },
          "predicted_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicted Tier"
          },
          "prediction_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prediction Confidence"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/RecommendationBlock"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "input_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Summary"
          },
          "creative_features": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creative Features"
          },
          "model_feature_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Feature Snapshot"
          },
          "prediction_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prediction Id",
            "description": "Row id after preview upserts PostgreSQL predictions (one row per campaign + metric)."
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "model_version",
          "campaign_intent",
          "target_metric",
          "target_label",
          "recommendations",
          "input_summary"
        ],
        "title": "PredictionPreviewResponse"
      },
      "PredictionResponse": {
        "properties": {
          "campaign_id": {
            "type": "integer",
            "title": "Campaign Id"
          },
          "predicted_ctr": {
            "type": "number",
            "title": "Predicted Ctr"
          },
          "predicted_conversion_rate": {
            "type": "number",
            "title": "Predicted Conversion Rate"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "campaign_id",
          "predicted_ctr",
          "predicted_conversion_rate",
          "id"
        ],
        "title": "PredictionResponse"
      },
      "PredictionRunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "campaign_intent": {
            "type": "string",
            "title": "Campaign Intent"
          },
          "target_metric": {
            "type": "string",
            "title": "Target Metric"
          },
          "target_label": {
            "type": "string",
            "title": "Target Label"
          },
          "predicted_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicted Tier"
          },
          "prediction_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prediction Confidence"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/RecommendationBlock"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "input_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Summary"
          },
          "creative_features": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creative Features"
          },
          "model_feature_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Feature Snapshot"
          },
          "prediction_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prediction Id",
            "description": "Row id after preview upserts PostgreSQL predictions (one row per campaign + metric)."
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "model_version",
          "campaign_intent",
          "target_metric",
          "target_label",
          "recommendations",
          "input_summary"
        ],
        "title": "PredictionRunResponse"
      },
      "RecommendationBlock": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "primary_kpi": {
            "type": "string",
            "title": "Primary Kpi"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "hint": {
            "type": "string",
            "title": "Hint"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "primary_kpi",
          "score",
          "hint"
        ],
        "title": "RecommendationBlock"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}